simplify the random system
