+++ RandomGen/IRandomSystem.cs
@@ -1,0 +1,7 @@
+﻿namespace RandomGen
+{
+    public interface IRandomSystem
+    {
+        double NextValue();
+    }
+}
